user情報更新のバリデーション

+++ myPorto/database/migrations/2014_10_12_000000_create_users_table.php
@@ -21,7 +21,7 @@ class CreateUsersTable extends Migration
             $table->string('name_kana')->nullable()->comment('ユーザー名カナ');
             $table->string('phone_number')->nullable()->comment('電話番号');
             $table->string('shipping_address')->nullable()->comment('配送先住所');
-            $table->integer('card_number')->nullable()->comment('クレジットカード番号');
+            $table->bigInteger('card_number')->nullable()->comment('クレジットカード番号');
             $table->string('card_name')->nullable()->comment('カード名義');
             $table->integer('expiration_date')->nullable()->comment('有効期限');
             $table->integer('cvc')->nullable()->comment('cvc');
